myStudy_Practice/20260209/my/study01/app1/pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.5",
    "jose>=1.0.0",
    "kafka-python>=2.3.0",
    "mariadb>=1.1.14",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "redis>=7.1.0",
]
